+++ packages/adhere/es/iframe-i-o/index.js
@@ -1,0 +1,2 @@
+import Model from"@baifendian/adhere-util-iframeio/es";export default Model;
+//# sourceMappingURL=index.js.map