+++ packages/adhere/src/index.ts
@@ -61,6 +61,7 @@ import ReactUtil from '@baifendian/adhere-util-reactutil';
 import Resource from '@baifendian/adhere-util-resource';
 import Validator from '@baifendian/adhere-util-validator';
 import WatchMemoized from '@baifendian/adhere-util-watchmemoized';
+import IframeIO from '@baifendian/adhere-util-iframeio';
 
 export { 
   BackTopAnimation,
@@ -126,4 +127,5 @@ export {
   Resource,
   Validator,
   WatchMemoized,
+  IframeIO,
 };